Big image full support!
It's here!
It's been requested, so here it is.
Click the "Toggle Long Images" button to get a long main image, and long
right image.

They auto crop to the length of the background! No more black space at
the bottom.

Also changed the names of saved images to be more user friendly.

Removed some residual code.